api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.2.1",
    "bson>=0.5.10",
    "fastapi>=0.115.6",
    "logfire>=3.1.1",
    "orjson>=3.10.14",
    "pynacl>=1.5.0",
    "redis-om>=0.3.3",
    "uvicorn>=0.34.0",
    "uvloop>=0.21.0",
]

[dependency-groups]
dev = [
    "pymongo>=4.10.1",
    "tqdm>=4.67.1",
]
